Refactor artefact and dashboard filter views
